requirements.txt
kafka-python
minio
fsspec
